feat(pod-template.tp): explicitly disable syscall source when driver is disabled
Starting from version 0.33.00, Falco supports multi sourcing. More than one source
can be enabled at the same time. By default the syscall source is always enabled.
When a new plugin is configured Falco automatically enables the source for that pluging.
If we do not want the syscall sourc then we need to disable it using the "--disable-source"
flag. This commit does that when the driver has no been enabled.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>