update!:(falco/driver-loader): simplify driver-loader logic
In order to deploy the falco-driver-loader init container there were two switches:
  * driver.loader.enabled=true
  * driver.loader.initContainer.enabled=true

There switche were needed to support three cases:
  1. driver loader via an init container: both options must be enabled;
  2. driver loader in the same falco image (with no init container): the first option must be enabled, the initContainer must not;
  3. the user provides the driver by other means (eg. install it directly in the host): both options must be disabled, but the driver feature is still enabled.

This commit removes the need for the driver.loader.initContainer.enabled=true. By doing so we still support case 1 and case 3 but not case 2.
Case 2 was needed when using the "falcosecurity/falco" docker image that contains the "falco-driver-loader" bash script.
Starting from version 2.0.0 of the charts the falcosecurity/falco-no-driver is the default image. Since version 2.2.0 of the charts the "falcosecurity/falco"
does not work anymore. Here we drop the support for it, for better support on the init container approach.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>