update(falco/templates): add `HOST_ROOT` env
Although HOST_ROOT is already set in all container images consumed by this chart, its default value (i.e. `/host`) is hard-coded in many points across this chart. So for consistency, we also force set it in `env:` for the main container and the initContainer as well.

The alternative would be to make it parametric, but since this is just an implementation detail that does not produce a user-facing effect, there's no compelling reason for the user to modify it.  Moreover, the hard-coded value works since its usage is consistent across and limited to only containers managed by this chart.

Signed-off-by: Leonardo Grasso <me@leonardograsso.com>